+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: valerio <valerio@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/24 19:01:21 by valerio           #+#    #+#             */
-/*   Updated: 2024/02/28 17:17:46 by valerio          ###   ########.fr       */
+/*   Updated: 2024/03/02 18:06:06 by valerio          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ int		ft_isalpha(int c);
 int		ft_isdigit(int c);
 int		ft_isalnum(int c);
 int     ft_isascii(int c);
-int     ft_strlen(const char  *s);
+size_t  ft_strlen(const char  *s);
 int		ft_isprint(int c);
 void    ft_bzero(void *s, size_t n);
 void    *ft_memcpy(void *dest, const void *src, size_t n);
